+++ audionimbus/src/effect/pathing.rs
@@ -411,7 +411,7 @@ pub struct PathEffectParams {
 
     /// Ambisonic coefficients for modeling the directional distribution of sound reaching the listener.
     /// The coefficients are specified in world-space, and must be rotated to match the listener’s orientation separately.
-    pub sh_coeffs: ShCoeffs,
+    pub sh_coeffs: Vec<f32>,
 
     /// Ambisonic order of the output buffer.
     /// May be less than the maximum order specified when creating the effect, in which case higher-order [`Self::sh_coeffs`] will be ignored, and CPU usage will be reduced.
@@ -435,23 +435,18 @@ pub struct PathEffectParams {
     pub normalize_eq: bool,
 }
 
-/// The spherical harmonic coefficients used in [`PathEffectParams`].
-/// Do not access these pointers after applying the effect.
-#[derive(Debug)]
-pub struct ShCoeffs(pub *mut f32);
-
-unsafe impl Send for ShCoeffs {}
-
-impl ShCoeffs {
-    pub const fn raw_ptr(&self) -> *mut f32 {
-        self.0
-    }
-}
 impl From<audionimbus_sys::IPLPathEffectParams> for PathEffectParams {
     fn from(params: audionimbus_sys::IPLPathEffectParams) -> Self {
+        let num_coeffs = (params.order + 1).pow(2) as usize;
+        let sh_coeffs = if params.shCoeffs.is_null() {
+            vec![]
+        } else {
+            unsafe { std::slice::from_raw_parts(params.shCoeffs, num_coeffs).to_vec() }
+        };
+
         Self {
             eq_coeffs: params.eqCoeffs,
-            sh_coeffs: ShCoeffs(params.shCoeffs),
+            sh_coeffs,
             order: params.order as u32,
             binaural: params.binaural == audionimbus_sys::IPLbool::IPL_TRUE,
             hrtf: params.hrtf.into(),
@@ -465,7 +460,7 @@ impl PathEffectParams {
     pub(crate) fn as_ffi(&self) -> FFIWrapper<'_, audionimbus_sys::IPLPathEffectParams, Self> {
         let path_effect_params = audionimbus_sys::IPLPathEffectParams {
             eqCoeffs: self.eq_coeffs,
-            shCoeffs: self.sh_coeffs.raw_ptr(),
+            shCoeffs: self.sh_coeffs.as_ptr().cast_mut(),
             order: self.order as i32,
             binaural: if self.binaural {
                 audionimbus_sys::IPLbool::IPL_TRUE
@@ -522,10 +517,8 @@ mod tests {
             let hrtf_settings = HrtfSettings::default();
             let hrtf = Hrtf::try_new(&context, &audio_settings, &hrtf_settings).unwrap();
 
-            let mut sh_storage = vec![0.0f32; NUM_SH_COEFFS];
-            sh_storage[0] = 1.0;
-
-            let sh_coeffs = ShCoeffs(sh_storage.as_mut_ptr());
+            let mut sh_coeffs = vec![0.0f32; NUM_SH_COEFFS];
+            sh_coeffs[0] = 1.0;
 
             let path_effect_params = PathEffectParams {
                 eq_coeffs: [1.0, 1.0, 1.0],
@@ -576,10 +569,8 @@ mod tests {
             let hrtf_settings = HrtfSettings::default();
             let hrtf = Hrtf::try_new(&context, &audio_settings, &hrtf_settings).unwrap();
 
-            let mut sh_storage = vec![0.0f32; NUM_SH_COEFFS];
-            sh_storage[0] = 1.0;
-
-            let sh_coeffs = ShCoeffs(sh_storage.as_mut_ptr());
+            let mut sh_coeffs = vec![0.0f32; NUM_SH_COEFFS];
+            sh_coeffs[0] = 1.0;
 
             let path_effect_params = PathEffectParams {
                 eq_coeffs: [1.0, 1.0, 1.0],
@@ -629,10 +620,8 @@ mod tests {
             let hrtf_settings = HrtfSettings::default();
             let hrtf = Hrtf::try_new(&context, &audio_settings, &hrtf_settings).unwrap();
 
-            let mut sh_storage = vec![0.0f32; NUM_SH_COEFFS];
-            sh_storage[0] = 1.0;
-
-            let sh_coeffs = ShCoeffs(sh_storage.as_mut_ptr());
+            let mut sh_coeffs = vec![0.0f32; NUM_SH_COEFFS];
+            sh_coeffs[0] = 1.0;
 
             let path_effect_params = PathEffectParams {
                 eq_coeffs: [1.0, 1.0, 1.0],